Thaiaddress/requirements.txt
joblib==1.4.2
pandas==1.5.3
numpy==1.26.4
fuzzywuzzy==0.18.0
spacy==3.7.4
python-crfsuite>=0.9,<0.10
pythainlp==5.0.2
regex==2023.12.25
Levenshtein==0.25.1
sklearn_crfsuite==0.5.0
jsonlines==4.0.0
fastapi==0.114.2
uvicorn
transformers==4.39.3
sentencepiece==0.2.0
sqlalchemy==2.0.34
psycopg2-binary==2.9.9
phonenumbers==8.13.46
python-dotenv
